backend/requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
boto3==1.42.96
botocore==1.42.96
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
colorama==0.4.6
cryptography==47.0.0
Django==6.0.4
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.2
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
googleapis-common-protos==1.74.0
google-genai
grpcio==1.80.0
grpcio-status==1.71.2
gunicorn==25.3.0
httplib2==0.31.2
idna==3.13
jmespath==1.1.0
mysqlclient==2.2.8
packaging==26.2
pillow==12.2.0
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
PyJWT==2.12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
redis==7.4.0
requests==2.33.1
s3transfer==0.16.1
six==1.17.0
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
uritemplate==4.2.0
urllib3==2.6.3
